moved market card out of markets-view
